Initial import of version V0.3 from Biela.
Version V0.3 was exported from Biela, because it became obvious that it is not meaningful to create script and try to fix problems with ChatGPT for Biela, when ChatGPT does not have access to Biela and does not know how things were implemented in Biela. This creates a lot of extra work and requires many tokens to fix issues.